Task 5 - Refactor auth service
