+++ .gitignore
@@ -1,1 +1,2 @@
 .claude/
+.vscode/
