+++ auth-service/src/routes/login.rs
@@ -1,0 +1,5 @@
+use axum::{response::IntoResponse, http::StatusCode};
+
+pub async fn login() -> impl IntoResponse {
+    StatusCode::OK.into_response()
+}
