+++ auth-service/src/routes/logout.rs
@@ -1,0 +1,5 @@
+use axum::{response::IntoResponse, http::StatusCode};
+
+pub async fn logout() -> impl IntoResponse {
+    StatusCode::OK.into_response()
+}
