+++ auth-service/src/routes/mod.rs
@@ -1,0 +1,12 @@
+mod login;
+mod logout;
+mod signup;
+mod verify_2fa;
+mod verify_token;
+
+// re-export items from sub-modules
+pub use login::*;
+pub use logout::*;
+pub use signup::*;
+pub use verify_2fa::*;
+pub use verify_token::*;
