+++ auth-service/src/routes/signup.rs
@@ -1,0 +1,5 @@
+use axum::{response::IntoResponse, http::StatusCode};
+
+pub async fn signup() -> impl IntoResponse {
+    StatusCode::OK.into_response()
+}
