+++ auth-service/src/routes/verify_2fa.rs
@@ -1,0 +1,5 @@
+use axum::{response::IntoResponse, http::StatusCode};
+
+pub async fn verify_2fa() -> impl IntoResponse {
+    StatusCode::OK.into_response()
+}
