+++ auth-service/src/routes/verify_token.rs
@@ -1,0 +1,5 @@
+use axum::{response::IntoResponse, http::StatusCode};
+
+pub async fn verify_token() -> impl IntoResponse {
+    StatusCode::OK.into_response()
+}
