+++ auth-service/tests/api/signup.rs
@@ -1,0 +1,10 @@
+use crate::helpers::TestApp;
+
+#[tokio::test]
+async fn signup_returns_200() {
+    let app = TestApp::new().await;
+
+    let response = app.post_signup().await;
+
+    assert_eq!(response.status().as_u16(), 200);
+}
